#82 Pikme 클릭 시 Piki 추가되는 로직 구현

+++ core_network/src/main/java/com/jyp/core_network/jyp/model/Piki.kt
@@ -11,7 +11,7 @@ data class Piki(
     val name: String,
     val address: String,
     val category: PlaceCategory,
-    val longitude: Long,
-    val latitude: Long,
+    val longitude: Double,
+    val latitude: Double,
     val link: String,
 )
